chore: addressed comments
